Implement instruction coverage tool
This is a timple tweak to the risc_runner to see how many of the
instructions in my test program are being executed. The basic design is
very simple. Because programs are relatively small, the interpreter
keeps a bitset of PC values, and sets ever time an instruction runs.
When the interpreter halts, it prints the PC of any instructions not
run.

This includes a tweaked version of ShortCircuit, which now executes all
the if conditional branches.