Prepare instructions for CASE statement
CASE statements will be implemented via a jump table, where the case
item will jump to the statement associated with that label.

This involves two new instructions, neither of which are "real"
instructions. The first is LDWC, which is the same as LDW but reads from
the code "segment" rather than from RAM. In a real CPU this wouldn't be
necessary becasue code and data would both be in RAM. However, in this
architecture it is nice because code can't be tampered with. The second
is CJP, which isn't an instruction at all. In a real CPU, this would be
a literal number embedded in the code stream, indicating the body of one
of the case items. Right now the VM doesn't support literal values, so
CJP is a placeholder Instruction, where the offset will have the PC to
be jumped to.