Generate code for CASE statements
CASE statements are conceptually syntactic sugar for IF/THEN, but can be
implemented compactly as a jump table. Each destination integer
represents an entry in the jump table, from smallest to highest. Jumping
to a case therefore means computing an offset into the table and then
jumping there.  For example, the example

  CASE a OF
    0 :WriteInt(100)
  | 1 :WriteInt(101)
  | 2 :WriteInt(102)
  | 3 :WriteInt(103)
  | 4 :WriteInt(104)
  | 5 :WriteInt(105)
  | 8 .. 10 :WriteInt(3333)
  END;

Generates the following code:

11	LDW R0, SP, #4
12	CMP R0, #10	; Jump over case if out of bounds
13	BGT 57
14	CMP R0, #0	; Jump over case if out of bounds
15	BLT 57
16	ADD R0, R0, #18	; Offset into jump table
17	B R0		; Jump to one of the values below
18	CASE 29
19	CASE 33
20	CASE 37
21	CASE 41
22	CASE 45
23	CASE 49
24	CASE 57		; Hole in case table goes after CASE
25	CASE 57
26	CASE 53
27	CASE 53
28	CASE 53
29	MOV R0, #100
30	MOV R1, #0
31	STW R0, R1, #-1
32	B 57
33	MOV R0, #101
34	MOV R1, #0
35	STW R0, R1, #-1
36	B 57
37	MOV R0, #102
38	MOV R1, #0
39	STW R0, R1, #-1
40	B 57
41	MOV R0, #103
42	MOV R1, #0
43	STW R0, R1, #-1
44	B 57
45	MOV R0, #104
46	MOV R1, #0
47	STW R0, R1, #-1
48	B 57
49	MOV R0, #105
50	MOV R1, #0
51	STW R0, R1, #-1
52	B 57
53	MOV R0, #3333
54	MOV R1, #0
55	STW R0, R1, #-1
56	B 57
57	MOV R0, #0	; Code after

Note that there "CASE pc" is a pseudo instruction which is simply
unconditional branch, "B pc". Thus there are two branches. One whch is
an indirect branch to the case table, and the second, in the case table,
to that paricular case. Thus, the code is
16	ADD R0, R0, #18	; Offset into jump table
17	B R0		; Jump to one of the values below
...
20	CASE 37

But potentially could be
  ADD R0, R0, #18  ; R0 contains an offset in the case table
  LDW R0, R0, 0    ; R0 now contains the PC to jump to
  B R0             ; Branch there

It turns out that in the current architecture, code and data don't share
an address space, so some contortions would be needed to do this (hence
the LDWC instruciton added in a previous commit which is now no longer
needed.)  It is unclear whether in practice a load followed by a jump
would be much slower than two jumps (probably would only be a pipeline
fail), but either would be possible.