Generate code for static array copies
With this code, a copy from one statically sized array to another
compiles and generates the correct code. With this, ArTest.ob compiles
with the RISC code generator.

This also generates code that wasn't supported in C, allowing the
compilation of copies of arrays of records (C codegen only allows string
copies)