Add a string copy from project Oberon.
Included is the source and the compiled RISC code.

An interesting thing to note here is that if the string being copied
does not fit in the destination string, it does *NOT* copy but instead
executes a TRAP. So the string needs to fit in the array, including
the final null termination byte.