Fix compiler code when optimization is turned on
calloc has been reverted because the issue wasn't pointers, but
something far more interesting.

With gcc -O2 on a Raspberry Pi, the core problem seemed to be that
preforming the operation

  hash := hash * 16777619;

performs very differntly with optimizations turned on and off. With
optimizations turned off, this seems to do what I expected, which was a
32bit x 32bit multiplication, and some overflow handling. With
optimizations turned off, this instead seemed to rather consistently
return INT_MIN (-2147483648) which is a reasonable answer considering
there are likely underflows, but isn't what I was expecting.

The end result was that the bucket being chosen was *always* 0, turning
the hash lookup into a linear search, and ultimately failing. There is
likely a subtle bug there as well, since the hash insert should be
robust to this, but the behavior would have been terrible because it
would have amounted to n^2 insertion performance.

The fix is to ensure that the number prior to the multiplication is
always positive, which seems to lead to more predictable multiplications
with optimizations turned on or off.

This may not be the correct way to implement the FNV hash, but at least
with this change, the code can be compiled with -O2, builds a working
compiler, and doesn't have broken lookup performance.

The proper fix would be to guarantee 32 bit integer multiplication and
addition operations, but getting the C compiler to do that can be
tricky.