Add a new test
This test computes the knights tour of a chessboard. A C version is
included as a comparison implementation, and to test correectness.

This produces the correct output with the C code generator, but right
now seems to hang on the RISC compiler, so the heavy recursion and / or
code generation may be an issue.