Add a pentominoes solver
This is another example of a solver by recursive descent. It solves
pentomino puzzles, which involves using the 12 pentomino shapes to fill a
board of various dimensions.

A large block of the solver is based on code for the 63 possible
pentomino shapes (e.g., when a pentomino can be flipped or rotated).
This information was generated by a Python script which took shape
coordinates as input and created all the distinct variants.