Transform run-tests.sh into a regression test suite.
Now, run-tests does not rebuild golden output, but instead builds each
test, runs them, and then compares the run output to the golden output.
If there is a diff, the script prints the diff and exits out. It's
simple, but it (basically) works.

Note that one test (IOTest) requires input, so stdin is given a golden
input that is passed to the tests.