Reorganize tests based on compiler.
Some of the test classes require C++ (because they rely on type
extension functionality or RTTI), whereas others can work with the
straight C compiler.

This reorganizes the tests such that the C++ tests use the -cpp option,
and the pure C ones do not. This exercises both compiler branches and
avoids slightly more expensive C++ compilation when not necessary.