Fix copying strings to array of char
ArrayTest was broken because when the string pool was being written, the
beginning of a string wasn't necessarily on a word boundary, which broke
structure copies. Now, static strings are padded to four byte
boundaries.

Note that, in the current code generation, it is possible (likely) that
too much is copied. The reason for this is because the structure copy
code looks at the size of the LHS (the fixed sized array). So, if the
destination array is 10 bytes, but the (static) string is less, it still
copies 12 bytes. Strictly speaking, this probably isn't an error,
because the contents of the invalid characters after a string are
undefined, but it is wasteful since it copies more than it should. It
can also potentially copy things after the end of the string pool which
is undesirable.

Copying where the source is an open array likely still does not work.