Add tests for copying an open array
This now should work properly in the C code generator and the risc
compiler.

It is very important to note that this behaves like the other static
copy examples. Since the destination size is fixed, that much data is
copied from the source.

This could cause a major problem if the source does not have as much
data as the destination, as the source will be read, followed by any
data that happens to be in memory after the source. This could
potentially cause problems with a large desintation but a small source,
where the small source is near the edge of memory (like the top of the
stack.)

The more careful solution needs to be to copy the minimum of the source
or the desintation, and do so dynamically. That code is more complex and
will come (hopefully) in the future.