Fix more array reference bugs
These bugs (with test cases added to ArrayTest) manifested themselves
when assigning from open arrays to reference parameters. The root cause
seems to be that the reference to the LHS of an assignment is sometimes
produced after the RHS. The end result was that the address to which a
write would happen was created and then overwritten, leading to memory
being written and then corrupted.