Proper fix for array / string copies
In previous versions, the number of bytes copied when copying a
structure were dictated by the LHS. In the case of strings (and open
array of chars) this can be problematic because the source must be
smaller than the destination, which means corrupt data would be copied
into the destination.

Now, the code generator properly generates bounds checks, so that if the
destination is bigger than the source, an error is raised, and only the
number of bytes of the destination are copied. This ensures that all the
data that is copied into the destination string is valid.