Fix offset bug with address lookup.
This bug was observed generating test code, but was difficult to put
into a test case. It has to do with array copying, where the copy
structure code was loading the pointer to the first array locaiton, but
then re-using the item. The end result was that the offset to the array
from the base pointer was used twice.