Fix case statement generation
When CASE statements had a gap in labels and the whole case statement
was offset (a.g., 'A'..'Z', 'a'..'z') the items not covered by the CASE
statement should simply skip past the statement. This was almost true
but was writing to the wrong place in the code buffer to fill in the
missing cases. With this, a CASE statement (similiar to one found in the
lexer) now passes tessts.