Implement Out.ReadFile in RISC interpreter
This allows a file from the filesystem to be inserted into the RAM of
the RISC virtual machine. There aren't many guard rails, but it works
for simple examples.

With this change, the RISC code generator and interpreter was able to
pass tests for its first and largest piece of non-trivial code. With
minor modifications (REAL numbers are not yet supported), the unit tests
of the Lexer (compiler/Lex.ob) pass when compiled and interpreted by the
RISC codegen.