Support REAL numbers in RISC code generation
This introduces several new opcodes for dealing with floating point
numbers. The code generator relies on the fact that a single precision
floating point number is 32 bits, and thus fits in a single register.
Because the code generator constants generally deal with integers, Out
now has a method that takes the bits of a REAL number and interprets
them as an integer. This could be avoided if the assembler outputted
floats, but this was a simpler solution.

The interpreter converts the bits of a register back and forth to a
float representation, based on the assumption that sizeof(int32_t) ==
sizeof(float). Otherwise, this is probably OK because a regier always
contains either an INTEGER or a REAL.

The tests added include a very primitive Mandelbrot set generator, as
well as some stress tests of real number instructions.