Re-enable bounds checks
This was an experiment to see how expensive bounds checking was. The
Knights tour in the RISC runner was running quite slowly, partly because
of the extreme number of array references.

Bounds checks are helpful, but for intense array references they do add
up.

On a Raspberry Pi 4, time to run KnightsTour without bounds checks:
real    1m33.844s

with bounds checks:
real    1m51.448s