Implement ORD and ABS code generation
These two builtins augment text and number processing. ORD converts a
char to an integer, which is an identity given the register
representatino. ABS is a small snippet of code to ensure a REAL is > 0,
and if not multiplies it by -1.

This change includes some new test cases, but more importantly, with
this change, the Lex.ob self-tests run correctly without any
modifications.