Remove Mute and Unmute from Out
The native Mute and Unmute functions were originally added because some
of the self-tests intentionally cause errors that the tests are
intenting to catch. This led to weird output, but with the new automated
run-tests, having the expected errors are OK.

With this change, The RiSC compiler can successfully run the Lexer
self-tests without modifications.