Add a proper lexer test
The lexer self-test is exported, so it can be trivially called by a test
function. This has the nice side-effect of the compiler importing Lex,
which has as non-trivial module initializer, which works properly.

Some minor tweaks were necessary to the test runner, because the lex
tester works by reading Lex.ob into memory and lexing it. This didn't
work when run from the test directory, but works if it runs directly
from the build directory.