Add support for NEW builtin
With this change, the RISC code generator support dynamic memory
allocation. This means supporting NEW(p) where p is a pointer.

The dynamic memory allocator is the simplest allocation imaginable. A
global variable is reserved at GB+0 which is a pointer to the next
available location where memory can be allocated (the start of the
"heap"). this is initialied at startup to be directly after all other
global variables and the string pool, so it is at the bottom of memory
and grows upward (the stack was already initialied to the top of memory
and grows downward)

To allocate memory, the heap address is assigned to the pointer, and the
heap pointer is increased by the size of the structure being allocated.
No other checking or initialization is performed. With this simplistic
approach, there is no garbage collection; all new objects stay in the
heap forever (based on the assumption that most programs are
short-lived).

This change is accompanied by a simple test case, but with this change,
much more complicated code now works that does pointer chasing, such as
the Decls self-tests in the compiler.