(Broken) experimental code demonstrating RETURN breakage
This isolates a bug that generates correct code in the C codegen, but

PROCEDURE BoolTest(i :INTEGER) :BOOLEAN;
BEGIN
  RETURN IsIOne(i) OR IsITwo(i) OR IsIThree(i)
END BoolTest;

The generated RISC code for which is this:

TestBed_BoolTest: code offset: 36
	; i offset 4
36	SUB SP, SP, #8
37	STW LR, SP, #0
38	STW R0, SP, #4
39	LDW R0, SP, #4
40	BL 0 (TestBed_IsIOne)
41	CMP R0, #0
42	BNE 36
43	LDW R0, SP, #4
44	BL 12 (TestBed_IsITwo)
45	CMP R0, #0
46	BNE 42
47	LDW R0, SP, #4
48	BL 24 (TestBed_IsIThree)
49	LDW LR, SP, #0
50	ADD SP, SP, #8
51	B LR

Note the code at line 42, which jumps back to 36, because the trueChain
and falseChain are not properly initialized.

This will be fixed in a followup commit, but this intermediate state was
useful for debugging, so I decided to snapshot here before making code
generation changes.