Fix generation of short circut boolean return
This fixes the problem described in commit e9671f0 and reintroduces the
full TestBed suite.

The core problem was that code like
  RETURN (a = 1) OR (a = 2)

would generate proper code, becuase the item was a conditional, and thus
LoadValue properly generated a proper boolean return value (0 or 1)

If, however, this code was
  RETURN (a = 1) OR OtherFunc(a)

then the item was the last code generated, in this case a procedure
call. In that case, the trueChain and falseChain were never fixed,
because LoadValue was unaware that fixing them was necessary.

The fix here appears to work, at the expense of slightly redundant code:

PROCEDURE BoolTest(i :INTEGER) :BOOLEAN;
BEGIN
  RETURN IsIOne(i) OR IsITwo(i) OR IsIThree(i)
END BoolTest;

Now generates the working code

TestBed_BoolTest: code offset: 519
	; i offset 4
519	SUB SP, SP, #8
520	STW LR, SP, #0
521	STW R0, SP, #4
522	LDW R0, SP, #4
523	BL 483 (TestBed_IsIOne)
524	CMP R0, #0
525	BNE 536
526	LDW R0, SP, #4
527	BL 495 (TestBed_IsITwo)
528	CMP R0, #0
529	BNE 536
530	LDW R0, SP, #4
531	BL 507 (TestBed_IsIThree)
532	CMP R0, #0
533	BNE 536
534	MOV R0, #0
535	B 537
536	MOV R0, #1
537	LDW LR, SP, #0
538	ADD SP, SP, #8
539	B LR

In this case, the code at line 536 is strictly not necessary, because
the boolean functions put the proper value in R0, but this could be
neccesaary in other conditions, and has the nice side effect of properly
generating the fixed trueChain and falseChain.