Make RISC interpreter memory size configurable
For most use cases, the default memory size of 64K words (256K bytes) is
sufficient and reasonably compact. In many examples, smaller RAM sizes
would suffice, and in others, like the compiler itself, more RAM is
necessary. Now, the memory size can be configured as an argument to
rcomplie, for example

 MEM_SIZE=262144 ../build/rcompile TestCompiler.ob