Add test harness for self hosting compiler
The compiler-test.sh shell script compiles the compiler into RISC
assembly, and then the RISC interpreter can compile the compiler. If the
output is identical, this should exercise much of the code generation.

The compiler is thus completely self hosting! Theoretically, the C
bootstrap or code generator are no longer necessary; this compiler could
simply ship with the C interpreter and the RISC code for the compiler.
That said, the other variants of the compiler are very useful.