Support lowercase keyword names.
This extends the compiler so that Oberon's syntactic keywords (e.g,
"PROCEDURE", "BEGIN" can be entered either in upper or lower case. This
leads to a langugage that looks a lot more like Pascal (or other modern
languages) and is easiert to type without using the shift key.

Note however that identifiers are still (intentionally) case sensitive,
so ProcedureeName and procedurename are different. The motivation behind
this is that it is that capitalization rules out to be consistent and
convey information, (e.g., CONSTANT vs ProcedureName vs localVar). It
would seem odd to use them inconsistently.

The ramification of this is that the builtin types (like INTEGER and
REAL) must (currently) still appear in upper case.