Fix generation of types within procedures
This adds a test case demonstrating how Oberon can implement a visitor
pattern using callback functions. The callback function needs a state
that is local to the procedure. In the C code generator, forward
references like this would not have worked because the typedef for the
inner type declaration would have been generated before it is used.

Thus, like nested procedures, nested types are now promoted to top
level.