Improve C code generation of type extension
Previously, with RECORDS defined like this:
  R1 = RECORD a, b: INTEGER END;
  R2 = RECORD (R1) c, d: INTEGER END;

The generated C code would look like:
  struct R1 { int a; int b; }
  struct R2 { int a; int b; int c; int d; }

In C, these are different types, and there is no guarantee that it is
reasonable to cast a pointer to R2 to a pointer to R1 (though it always
seems to work.)

Now, the compiler generates different code for R2:
  struct R2 { struct R1 _base; int c; int d; }

According to (my intepretation of) the C99 specification, it should now
guarantee that the down cast of R2 to R1 should work, becuase (a) _base
is exactly an R1, and (b) C99 guarantees that the first element of a
struct be the first thing pointed to. This requires a cast to avoid
compiler warnings, but (hopefully) this is safe across C compilers and
architectures.

This means that references to things like r2.a in Oberon now look like
r2._base.a in C, but since this is a constant offset, the compiled code
ought to be exactly the same.

To accomplish this, the decl that represents a field re-uses the
"nesting" attribute, which previously was only used for variables and
procedures.  The fields c and d have a nesting level of 0, and a and b
(inside the definition of R2) have a nesting level of 1, giving the
C code generator the hint of adding the "_base" field references.

In C++, the compiler likely does a transformation like this
automatically, and thus doesn't need the extra casts or boilerplate.

With this change, the RecCopy example, which required the C++ compiler,
now runs properly with the C codegen without warning.