Unify C / C++ RECORD generation
The code paths for generating RECORDS in the C code generator had been
copied and pasted for simplicity. However, with the previous
rearchitecting in 51a42579, the code of the two were very similar.  Now,
in C++,
  R1 = RECORD a, b: INTEGER END;
  R2 = RECORD (R1) c, d: INTEGER END;

Generates code like this:
  struct R1 { virtual ~R1() {} int a; int b; }
  struct R2 : public R1 { int c; int d; }

This big difference between this and the previous incarnation was the
removal of the syntheic Oberon type id, for in C++ each record carried a
4 byte type id used for IS and typeguard testing. Instead, the new
version uses C++ RTTI, which likely makes use of the vtable which C++
was already adding. A typeguard
  r is R2
looks like this
  dynamic_cast<R2*>(%r) != nullptr

The C code still can't handle IS or typeguard, but now that the remnants
of the old oberon_typeinfo have been removed, that code can be replaced
(with something that will work in both C and ported to the RISC code
generator).