Keep track of the nesting level of a record
for a record R1 = RECORD END, the nesting level of R1 is 0. For
R2 = RECORD (R1) END, R2's nesting level is 1, for
R3 = RECORD (R2) END, R3's nesting level is 2. This information will be
helpful later for figuring out type extension.

Consder the following
  Base = RECORD END
  R1 = RECORD (Base) END    S1 = RECORD (Base) END
  R2 = RECORD (R1) END      S2 = RECORD (S1) END
  R3a = RECORD (R2) END
  R3b = RECORD (R2) END

Given Base, you can determine
  b IS R1 if the type id at nesting level 1 is R1. This is true of R1,
R2, R3a, and R3b, and not of S1 and S2.
  b IS R2 if the type id at nesting level 2 is R2. This is true of R2,
R3a, and R3b, but not any of the other types.