add a native code test
This is a more elaborate test of native code. In this case, the Cairo
graphics package is used to write a PNG. This defines a Cairo module for
interfacing with the Cairo graphics package, and exposes some of the
functions that Cairo exposes. The end test creates a cairo canvas, draws
a few lines, and writes a PNG.

This illustrates a flaw with the way that native code was
conceptualized. Here, the interface between the Cairo module and the
outside world is the "Cairo_Context" record, which is an opaque type as
far as callers are concerned.

In the current compiler imlementation, the C runtime in runtime.c is
imported first, and then cairo code. This leads to an unresolved forward
declaration of the context, which made the code a bit harder than it
should have been.