Rework how type extension tests work
This change dramatically alters how the compiler deals with type
descriptions and dynamic types.

The core of this change is the generation of "type descriptions".
Consider this example:

  R1 = RECORD a: INTEGER END;
  S1 = RECORD s: INTEGER; END;
  R2 = RECORD (R1) b: INTEGER END;
  R3 = RECORD (R2) c: INTEGER END;
  R4 = RECORD (R3) d: INTEGER END;

Each record that the compiler sees generates a type descriptor, which
looks like this:

oberon_typedesc oberon_typedescs[] = {
  { {0, -1, -1, -1, -1, -1, } }, // Entry 0 empty
  { {1, -1, -1, -1, -1, -1, } }, // Entry 1 for R1
  { {2, -1, -1, -1, -1, -1, } }, // Entry 2 for S1
  { {1, 3, -1, -1, -1, -1, } },  // Entry 3 for R2
  { {1, 3, 4, -1, -1, -1, } },   // Entry 4 for R3
  { {1, 3, 4, 5, -1, -1, } },    // Entry 5 for R4
};

Consider the last entry, the one for R4. The first element of that
is 1, indicating that R4 extends R1. The second entry is 3, indicating
that R4 extends R2, and so on.

The key advantage of this representation is that *any* RECORD that,
say, extends R2, will have the type id for R2 in its second slot. So
record X is an R2 if record X's second element of its type descriptor
has the type descriptor for R2.

In other words, a type check can (a) reference a static table, and
(b) determine if a type extends another in constant time, since it
looks for a static entry in a table, rather than chasing a parent
chain.

The other amazing thing about this approach is that only pointers
need a tiny amount of construction logic in their NEW method. When
passing a RECORD as a (VAR) parameter, it can pass a special variable
for its type. Much like __len is passed, for a record r, the
parameter r__typeid is also passed (only when needed). This can be
initialized as an integer constant, because when the variable for r
is defined, its (static) type id is known.

In C code, an IS test looks like this:

  a is R2

  generates

  (oberon_typedescs[a__typeid].parent[1] == 3)

For pointers, a little bookkeeping is necessary. Pointers have been
updated such that they are always at least 4 bytes bigger than they
need to be. The pointer still points to the record, but four bytes
before the pointer is the type of the pointer. Strictly speaking,
pointers could be the same as records, but this approach makes a lot
of sense with dynamic allocation, and is intended to be ported over
to the RISC code generator now that it works for C. So, the code

  p is PR2

  generates

  oberon_typedescs[*(((oberon_typeid*)p)-1)].parent[1] == 3)

The C pointer arithmetic is a little messy, because p gets
reinterpreted as a pointer to an int (typeid), -1 gets the memory just
before the pointer, and then that dereferenced.

So only pointers have a permanent runtime memory overhead for each
pointer (which they would have needed anyway for garbage collection.)

With this change, TypeExt.ob properly compiles and runs in the C code
generator. Some of the other tests don't work just yet because they
use dynamic_cast for typeguard, but changes for them should be
relatively minor.