Define behavior for NIL typeguard
The Oberon language specification explicitly states that the result of
NIL IS P is undefined. With this change, this complier ensures that the
result of that is FALSE. (Implementations shouldn't rely on this, but at
least for now this behavior is specific.)  This avoids potential
segmentation faults when attempting the type test, as well as avoiding
silly compiler warnings since a procedure like

PROCEDURE IsFoo(VAR f: Something): BOOLEAN;
BEGIN
  RETURN f IS Foo
END IsFoo;

really *does* use f, but in the C code it just uses the synthetic
f__typeid.