Generate type guards with C code generator
A type guard p(P) performs a run time check that p can be cast to P, and
then uses it. In C++, this used dynamic_cast. For C, this uses regular
casts, and a refactored version of the type test to perform the check
and cast.

The code:

  R1 = RECORD a: INTEGER END;
  R2 = RECORD (R1) b: INTEGER END;
  PR1 = POINTER TO R1;
  PR2 = POINTER TO R2;
  ...
VAR
  p :PR1;
BEGIN
  p(PR2).b := 3;

Is converted to

  (*(assert(p && oberon_typedescs[*(((oberon_typeid*)p)-1)].parent[1] == 4),
     ((PR2)(p)))).b = 3;

This mouthful uses the rarely used comma operator in C, (f1, f2) to run
two statements (the assert and the value) and return the second value.
The value could be cast directly, but the added type test provides
runtime type safety that the cast is valid.

With this change, all of the C++ specific tests now work properly in the
C compiler.