Update generated code to compile without warnings.
When assigning a pointer to a pointer whose type extends it (downcasting
the pointer), the semantic analysis has already confirmed that the
downcast is safe, so the C code can directly perform the cast (again,
this cast is unnecessary in C++).

In some of the example code, there was a "base" record which existed
merely as a sentinel at the top of the type tree. The C compiler warns
about this (which is probably legitimate in C, because the size of an
empty struct is unclear. It is possible that the code generator should
insert the dummy, rather than making the code do it, since what was
previously there is valid Oberon code.