Move dummy field creation into codegen.
Some C compilers warn when generating an empty struct, but this is
reasonable Oberon code, especially because an empty struct may be part
of a hierarchy of types.

In a previous commit, the test examples added this dummy field, but
now the code generator inserts the dummy field synthetically. The tests
have been returned to their previous (no field) state.