Fix bug when pointer dereferenced to record VAR param
This is a rather uncommon edge case, when you have a procedure
  PROCEDURE P(VAR r: R)

and caller
  VAR p: PR;
   ...
   P(p^)

There are two ways to get the type id for the derefenced pointer passed
into the function (because records that need a typeid get an invisible
parameter.) One way is to take the pointer value minus 4 bytes (the
dynamic type of the poitner). The other way is to return the record type
of the pointer (the static type.) For simplicity, this code does the
latter, which appears to be consistent with what Wirth's compiler does.

With this is added a new example of object oriented programming, which
uses dynamic dispatch but not dynamic memory allocaiton, by consistently
using VAR parameters (OopTest2.ob). As an example of mixing and matching
records and pointers, this bug was uncovered, which previously caused
the compiler to crash.