Add RISC code generation for type test
This adapts the approach prototyped in the C code generator to support
type identifiers in the RISC target. It adds support for the "IS" type
test, and paves the way for asserted type guards (right now, type guards
always succeed, but they should pass the runtime type test added in this
commit.)

The approach is to create a type table, which is a table of pointers
stored in low memory. The approach is discussed in detail in 1b90b861.
Unlike the C code, the type table entries are pointers to things in
memory, and looks like this:

  R1 = RECORD a: INTEGER END;
  S1 = RECORD s: INTEGER; END;
  R2 = RECORD (R1) b: INTEGER END;
  R3 = RECORD (R2) c: INTEGER END;
  R4 = RECORD (R3) d: INTEGER END;

causing this to be put into low memory;

const int32_t TYPE_TABLE[] = {
  4, -1, -1, -1, -1, -1,     // Entry 0 empty, starts at address 4
  28, -1, -1, -1, -1, -1,    // Entry for R1
  52, -1, -1, -1, -1, -1,    // Entry for S1
  28, 76, -1, -1, -1, -1,    // Entry for R2
  28, 76, 100, -1, -1, -1,   // Entry for R3
  28, 76, 100, 124, -1, -1,  // Entry for R4
};

The table begins at address 4 to leave room for the address 0, which is
the next heap byte to be allocated.

Since the type table entries are pointers, performing a type test for a
record is only two memory loads, e.g., "a is R3" compiles to

72	LDW R0, SP, #8		; Load typeinfo from the stack
73	LDW R0, R0, #8		; Get the entry for the type's "R3 slot"
74	CMP R0, #100		; Is it R3?
75	BEQ 78			; If so, do the true branch

Pointer comparisons are similary compact, but adding a NIL check, so "pa
IS PR3" compiles to

264	LDW R0, SP, #4		; Load poitner to pa
265	CMP R0, #0		; Is it NIL?
266	BEQ 271			; If so, branch to false
267	LDW R0, R0, #-4		; Pointer typinfo is pa-4 bytes
268	LDW R0, R0, #8		; Look up pointer's R3 slot
269	CMP R0, #100		; Is it R3?
270	BEQ 273			; If so, do the true branch

Much like the C code generator, RECORDS that need a typeinfo are passed
with an extra parameter (their type id), and pointers that need a
typeinfo are 4 bytes larger, the first 4 bytes being the type id.

With this change, the TypeExt.ob and OopTest.ob tests run properly in
the RISC runner.