Handle typeid of pointer dereference
Like the C code generator, the run time type of a pointer dereferene is
based on the type of that pointer. This appears to be consistent with
Wirth's compiler, and also with the language spec itself. That is, the
type of a dereferenced pointer is the type of thing it points to, not
necessarily the dynamic type of the object.

With this change, OopTest2 works properly in the generated RISC code