Simplify run-tests
With recent changes, nearly every test runs in every variation of the
compilers (C code generation, C++ code generation, RISC code
generation.) There are only a few exceptions.

This refactors the test harness to have a single list of all the tests,
with a few places where tests are excluded (and why).