Fix C code generation of nested procedure names
Previously, the C code generator was using a hack, turning the name of
a nested procedure into a different token. This approach usually doesn't
work right, and ends up attempting to generate a name for a function
called "TestBed_" or something similiar

Instead, this now does this the proper way, by generating a synthetic
name for the function and registering it with the lexer.  This ensures
that all inner procedures have guaranteed unique names, which is
necessary to avoid scope clash (since in C every function needs to be
delcared at top level.)

With this, TestBed now works for the C code generator, so there are no
exceptions needed for C anymore.