Rename LangExtensionsTest to LangExtensions
The "Test" suffix is redundant because everything in the tests directory
is a compiler test. It probably makes sense to rename more of these, but
LangExtensions was the most obvious one.

This is the second attempt at this rename, the first, incorrectly
updated the test script and failed to rename the goldens output.