Fix RISC code generation of nested procedures
With this update, the nested procedure tests, such as Recurse and
Recurse2, work properly in the RISC compiler (and indeed, now all three
compilers run nearly the same test suite!)

Tchis required a reworking of how nested procedures were compiled.
Previously, the tree shaker would treat each procedure as independent,
and send them in dependency order. This is still true for top level
procedures, but nested procedures could be inherently circular, such as

  PROCEDURE A;
    PROCEDURE B;
      BEGIN
        A;
      END B;
  END A;

To make this work, top level procedures are now responsibile for
generating any procedures defined within them. That way, the outer
procedure has a known starting address, which can be used by the innter
procedures.

This would have been considerably simpler had not there been an
assumption that codeBuffer be for the current procedure. That approach
doesn't work in a recursive procedure generator, so a little care was
needed to generate a procedure (and any nested procedures) all at once.
This seems a reasonable compromose since nested procedures are rare.