Implement string comparisons in RISC code generation
In Oberon, s1 = s2 on two strings (or arrays of characters) does a
proper string comparison. A string is an array of characters (bytes)
terminated by a NULL character. Two strings are equal if they are
byte-for-byte identical until the 0 (0X / NIL / NULL) byte. Strings can
also be compared with < or >, which performs byte comparison orderings.

The code pattern is relative straightforward, where s1 = s2 becoming

47	LDW R0, SP, #4		; Load s1
48	LDW R1, SP, #12		; Load s2
49	LDB R2, R0, #0		; Load *s1
50	LDB R3, R1, #0		; Load *s2
51	ADD R0, R0, #1		; s1++
52	ADD R1, R1, #1		; s2++
53	CMP R2, R3		; does *s1 == *s2 ?
54	BNE 57			; if not, done
55	CMP R2, #0		; if so, reached the end of string?
56	BNE 49			; if not, test again
57	BGE 62			; comparson after, in this case >=

This works because Oberon guarantees proper 0-terminated behavior  This
would fail for an ARRAY N OF CHAR that is not null-terminated, but that
behavior is undefined in Oberon.

With this change, the string tests now pass.

Also, with this change, The RISC code generator should be able to handle
all language features as described in the Oberon-07 report!