Add type tests to typeguards
The RISC code generator previously performed the cast associated with a
typeguard. It now asserts (at runtime) that the type guard is valid
before doing the cast. This causes an ASSERT failure if the type guard
is not for a vlid type. tests/errors/CastFail.ob validate this failure
behavior.