Add bounds check for static reference to open array
If the array size is known at compile time, the compiler can perform
bounds checks at compile time for static lookups, such as a[8]. It
cannot perform such checks for static lookups to open arrays.

These bounds checks were working in the C compiler, but were missing in
the RISC one, but now work in both.