Add a turtle graphics exaple
This example uses the sine table to add some commands for Turtle
Geometry, as popularied by the Logo programming langauge. A "turtle"
controls a pen as it moves around, with commands to move forward or turn
right by a certain number of degrees (negative movement goes backwards,
and negative angles turn left).

This example is personally meaningful because this is possibly one of
the earliest programs I ever wrote. We were taught Logo in Elementary
school, and one of the first programs I wrote was this:

  TO SQUARE
    REPEAT 4 [ FD 50 RT 90 ]
  END

  TO FLOWER
    REPEAT 20 [ SQUARE RT 18 ]
  END

The Oberon equivalent is very similiar. It all comes full circle.