Add memory corruptin detection to RISC interpeter
The memory limits on the interpreter were previously determined by
doubling memory until the RISC compiler binary ran without crashing. The
reason for the crashes was likely that the stack grew into the heap,
causing data to be overwritten.

The interpreter now does some rudimentry stack/heap collision detection
and halts immediately if the stack pointer goes lower than the heap
pointer, since the heap grows toward high memory and the stack grows
toward low memory.

This had the other advantage that we can now more precisely measure how
much memory a particular binary actually uses, by printing the stack
extent and heap use when the program exits. For posterity, this output
at the moment, when the compiler compiles itself, is:

Max heap 1930436 min stack 3638308 stack use 361692

So 1930436+361692 = 2292128 bytes, or 573032 4-byte words, which is how
the memory in the interpreter is allocated.